Fix bug in client metrics (#56)
This PR fixes #51 - the issue that happend is the below:

1. A testcase that generates new coverage but times out
2. The code grabbed the last new coverage, and added the amount of new coverage to the coverage counter
3. As the testcase timed-out, this coverage is asked to be revoked

Now let's say the same testcase runs again, retriggers this new coverage but times out; the amount of 'new' coverage gets added again even though it'll get reverted. This happening a large number of times lead to the coverage count being completely inacurrate on the client side.