Demonstrate multi-packets testcase delivery w/ fuzzer_tlv_server (#76)
This PR builds on top of the @y0ny0ns0n's work to show how a fuzzer module can deliver 'multi packets' testcases to a target. On top of that, it shows how the inputs can be stored as a human readable format and deserialized by the fuzzer module into structured data. (#69)