move the fs hooks from ntdll to nt for wow64 (#108)
This enables Wow64 snapshot to use the fs hooks. Having them at the `ntdll` layer meant that the CPU is still executing in 32-bit compatibility mode while executing the `Nt*` functions which means that the code for the hooks caused a bunch of issues: corrupting memory (32-bit vs 64-bit), truncations, etc.