Move the `human` functions in their own files / headers (#147)
This moves the `human` functions in their own headers and also simplify the way they can be used; they now return specific types that have their own `fmt::print` formatter which makes it easy to dump them on `stdout`. Prior, the caller needed to capture both the amount and the unit. This also fixes #145.