Fix memory performance issue related to `aligned_alloc` (#149)
I was profiling `wtf` and found a behavior that I wasn't aware of. Basically, every calls we make to `aligned_alloc` allocates 8kish instead of 4kish. For whatever reason, I made an assumption that it would use a special allocator; it isn't a special allocator it just wraps a call to `malloc` with that math: `v5 = malloc(Alignment - 1 + Size + 8);`. In our case, both `Alignment` & `Size` are `0x1000`.

I switched the calls to `aligned_alloc` to `VirtualAlloc` / `mmap` to match the behavior that I wanted in the first place. We also don't need page aligned chunks in the `Ram_t` object, so I turned this one into a `malloc` call.